W-14875628 (refactor): extracted each binding parsing behind an interface to be able to mix and match them
