+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/async/parser/bindings/AsyncServerBindingsParser.scala
@@ -7,6 +7,7 @@ import amf.apicontract.internal.metamodel.domain.bindings.{
   MqttServerLastWillModel,
   ServerBindingsModel
 }
+import amf.apicontract.internal.spec.async.parser.bindings.server.MqttServerBindingParser
 import amf.apicontract.internal.spec.async.parser.context.AsyncWebApiContext
 import amf.apicontract.internal.spec.common.WebApiDeclarations.ErrorServerBindings
 import amf.apicontract.internal.spec.spec.OasDefinitions
@@ -48,37 +49,6 @@ case class AsyncServerBindingsParser(entryLike: YMapEntryLike)(implicit ctx: Asy
   override protected def parseMqtt(entry: YMapEntry, parent: String)(implicit
       ctx: AsyncWebApiContext
   ): ServerBinding = {
-    val binding = MqttServerBinding(Annotations(entry))
-    val map     = entry.value.as[YMap]
-
-    map.key("clientId", MqttServerBindingModel.ClientId in binding)
-    map.key("cleanSession", MqttServerBindingModel.CleanSession in binding)
-    map.key("keepAlive", MqttServerBindingModel.KeepAlive in binding)
-
-    parseLastWill(binding, map)
-
-    parseBindingVersion(binding, MqttServerBindingModel.BindingVersion, map)
-
-    ctx.closedShape(binding, map, "mqttServerBinding")
-
-    binding
-  }
-
-  private def parseLastWill(binding: MqttServerBinding, map: YMap)(implicit ctx: AsyncWebApiContext): Unit = {
-    map.key(
-      "lastWill",
-      { entry =>
-        val lastWill    = MqttServerLastWill(Annotations(entry.value))
-        val lastWillMap = entry.value.as[YMap]
-
-        lastWillMap.key("topic", MqttServerLastWillModel.Topic in lastWill)
-        lastWillMap.key("qos", MqttServerLastWillModel.Qos in lastWill)
-        lastWillMap.key("retain", MqttServerLastWillModel.Retain in lastWill)
-        lastWillMap.key("message", MqttServerLastWillModel.Message in lastWill)
-
-        ctx.closedShape(lastWill, lastWillMap, "mqttServerLastWill")
-        binding.setWithoutId(MqttServerBindingModel.LastWill, lastWill, Annotations(entry))
-      }
-    )
+    MqttServerBindingParser.parse(entry, parent)
   }
 }
