+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/async/parser/bindings/operation/Amqp091OperationBindingParser.scala
@@ -1,0 +1,33 @@
+package amf.apicontract.internal.spec.async.parser.bindings.operation
+
+import amf.apicontract.client.scala.model.domain.bindings.amqp.Amqp091OperationBinding
+import amf.apicontract.internal.metamodel.domain.bindings.{Amqp091OperationBindingModel, KafkaOperationBindingModel}
+import amf.apicontract.internal.spec.async.parser.bindings.BindingParser
+import amf.apicontract.internal.spec.async.parser.context.AsyncWebApiContext
+import amf.core.internal.parser.YMapOps
+import amf.core.internal.parser.domain.Annotations
+import org.yaml.model.{YMap, YMapEntry}
+
+object Amqp091OperationBindingParser extends BindingParser[Amqp091OperationBinding] {
+  override def parse(entry: YMapEntry, parent: String)(implicit ctx: AsyncWebApiContext): Amqp091OperationBinding = {
+    val binding = Amqp091OperationBinding(Annotations(entry))
+    val map     = entry.value.as[YMap]
+
+    map.key("expiration", Amqp091OperationBindingModel.Expiration in binding)
+    map.key("userId", Amqp091OperationBindingModel.UserId in binding)
+    map.key("cc", Amqp091OperationBindingModel.CC in binding)
+    map.key("priority", Amqp091OperationBindingModel.Priority in binding)
+    map.key("deliveryMode", Amqp091OperationBindingModel.DeliveryMode in binding)
+    map.key("mandatory", Amqp091OperationBindingModel.Mandatory in binding)
+    map.key("bcc", Amqp091OperationBindingModel.BCC in binding)
+    map.key("replyTo", Amqp091OperationBindingModel.ReplyTo in binding)
+    map.key("timestamp", Amqp091OperationBindingModel.Timestamp in binding)
+    map.key("ack", Amqp091OperationBindingModel.Ack in binding)
+
+    parseBindingVersion(binding, KafkaOperationBindingModel.BindingVersion, map)
+
+    ctx.closedShape(binding, map, "amqpOperationBinding")
+
+    binding
+  }
+}
